Support for RecoveryServices Vault Upgrade- Taken comments
