+++ src/ServiceManagement/RecoveryServicesVaultUpgrade/VaultUpgrade.Tests/App.config
