+++ src/ServiceManagement/RecoveryServicesVaultUpgrade/VaultUpgrade.Tests/Helpers/RecordedDelegationHandler.cs
