+++ src/ServiceManagement/RecoveryServicesVaultUpgrade/VaultUpgrade.Tests/SessionRecords/VaultUpgrade.Tests.VaultUpgradeTests/VaultUpgrade_CheckPrereqsTests.json
@@ -1,0 +1,76 @@
+{
+  "Entries": [
+    {
+      "RequestUri": "/7a8c4b2b-1343-4ead-903f-031163054294/cloudservices/RecoveryServices-DKIEGNVV3OAPUDVOTINGUOLCIJIDA743FNSWQAM4O4NEQUVZFBWA-West-US/resources/HRMBvtd2ToDogfood/~/HyperVRecoveryManagerVault/TestVault2/checkPrereqsForUpgrade",
+      "EncodedRequestUri": "LzdhOGM0YjJiLTEzNDMtNGVhZC05MDNmLTAzMTE2MzA1NDI5NC9jbG91ZHNlcnZpY2VzL1JlY292ZXJ5U2VydmljZXMtREtJRUdOVlYzT0FQVURWT1RJTkdVT0xDSUpJREE3NDNGTlNXUUFNNE80TkVRVVZaRkJXQS1XZXN0LVVTL3Jlc291cmNlcy9IUk1CdnRkMlRvRG9nZm9vZC9+L0h5cGVyVlJlY292ZXJ5TWFuYWdlclZhdWx0L1Rlc3RWYXVsdDIvY2hlY2tQcmVyZXFzRm9yVXBncmFkZQ==",
+      "RequestMethod": "POST",
+      "RequestBody": "<ResourceUpgradeInput xmlns=\"http://schemas.microsoft.com/windowsazure\">\r\n  <NewResourcePath>/subscriptions/7a8c4b2b-1343-4ead-903f-031163054294/resourceGroups/TargetRg/providers/Microsoft.RecoveryServicesBVTD2/Vaults/TestVault2</NewResourcePath>\r\n</ResourceUpgradeInput>",
+      "RequestHeaders": {
+        "Content-Type": [
+          "application/xml"
+        ],
+        "Content-Length": [
+          "271"
+        ],
+        "Accept": [
+          "application/xml"
+        ],
+        "Agent-Authentication": [
+          ""
+        ],
+        "x-ms-client-request-id": [
+          "1fcc55a9-f6ee-4f72-90b5-749e9194f858"
+        ],
+        "x-ms-version": [
+          "2013-03-01"
+        ],
+        "User-Agent": [
+          "Microsoft.WindowsAzure.Management.RecoveryServicesVaultUpgrade.RecoveryServicesVaultUpgradeManagementClient/1.0.0.0"
+        ]
+      },
+      "ResponseBody": "<Error i:nil=\"true\" xmlns=\"http://schemas.microsoft.com/windowsazure\" xmlns:i=\"http://www.w3.org/2001/XMLSchema-instance\" />",
+      "ResponseHeaders": {
+        "Content-Length": [
+          "123"
+        ],
+        "Content-Type": [
+          "application/xml; charset=utf-8"
+        ],
+        "Expires": [
+          "-1"
+        ],
+        "Pragma": [
+          "no-cache"
+        ],
+        "x-ms-servedbyregion": [
+          "uswest"
+        ],
+        "x-ms-client-request-id": [
+          "1fcc55a9-f6ee-4f72-90b5-749e9194f858"
+        ],
+        "Strict-Transport-Security": [
+          "max-age=15724800; includeSubDomains"
+        ],
+        "x-ms-request-id": [
+          "0fe76d71ea8d7a9cb9a52be25b497897"
+        ],
+        "Cache-Control": [
+          "no-cache"
+        ],
+        "Date": [
+          "Thu, 23 Mar 2017 12:12:03 GMT"
+        ],
+        "Server": [
+          "1.0.6190.8255",
+          "(rd_rdfe_n.170217-1352)",
+          "Microsoft-HTTPAPI/2.0"
+        ]
+      },
+      "StatusCode": 200
+    }
+  ],
+  "Names": {},
+  "Variables": {
+    "SubscriptionId": "7a8c4b2b-1343-4ead-903f-031163054294"
+  }
+}
