+++ src/ServiceManagement/RecoveryServicesVaultUpgrade/VaultUpgrade.Tests/SessionRecords/VaultUpgrade.Tests.VaultUpgradeTests/VaultUpgrade_TrackResourceUpgradationTests.json
@@ -1,0 +1,70 @@
+{
+  "Entries": [
+    {
+      "RequestUri": "/7a8c4b2b-1343-4ead-903f-031163054294/cloudservices/RecoveryServices-DKIEGNVV3OAPUDVOTINGUOLCIJIDA743FNSWQAM4O4NEQUVZFBWA-West-US/resources/HRMBvtd2ToDogfood/~/HyperVRecoveryManagerVault/TestVault2/upgradeOperationResult",
+      "EncodedRequestUri": "LzdhOGM0YjJiLTEzNDMtNGVhZC05MDNmLTAzMTE2MzA1NDI5NC9jbG91ZHNlcnZpY2VzL1JlY292ZXJ5U2VydmljZXMtREtJRUdOVlYzT0FQVURWT1RJTkdVT0xDSUpJREE3NDNGTlNXUUFNNE80TkVRVVZaRkJXQS1XZXN0LVVTL3Jlc291cmNlcy9IUk1CdnRkMlRvRG9nZm9vZC9+L0h5cGVyVlJlY292ZXJ5TWFuYWdlclZhdWx0L1Rlc3RWYXVsdDIvdXBncmFkZU9wZXJhdGlvblJlc3VsdA==",
+      "RequestMethod": "GET",
+      "RequestBody": "",
+      "RequestHeaders": {
+        "Accept": [
+          "application/xml"
+        ],
+        "Agent-Authentication": [
+          ""
+        ],
+        "x-ms-client-request-id": [
+          "50c6d65d-361f-42b9-a2f9-64b46d03615b"
+        ],
+        "x-ms-version": [
+          "2013-03-01"
+        ],
+        "User-Agent": [
+          "Microsoft.WindowsAzure.Management.RecoveryServicesVaultUpgrade.RecoveryServicesVaultUpgradeManagementClient/1.0.0.0"
+        ]
+      },
+      "ResponseBody": "<OperationInfo xmlns=\"http://schemas.microsoft.com/windowsazure\" xmlns:i=\"http://www.w3.org/2001/XMLSchema-instance\">\r\n  <ErrorCode></ErrorCode>\r\n  <JobSteps i:nil=\"true\" />\r\n  <Jobs i:nil=\"true\" xmlns:d2p1=\"http://schemas.microsoft.com/2003/10/Serialization/Arrays\" />\r\n  <Message></Message>\r\n  <OperationResult>Succeeded</OperationResult>\r\n  <OperationStatus>Completed</OperationStatus>\r\n</OperationInfo>",
+      "ResponseHeaders": {
+        "Content-Length": [
+          "378"
+        ],
+        "Content-Type": [
+          "application/xml; charset=utf-8"
+        ],
+        "Expires": [
+          "-1"
+        ],
+        "Pragma": [
+          "no-cache"
+        ],
+        "x-ms-servedbyregion": [
+          "uswest"
+        ],
+        "x-ms-client-request-id": [
+          "50c6d65d-361f-42b9-a2f9-64b46d03615b"
+        ],
+        "Strict-Transport-Security": [
+          "max-age=15724800; includeSubDomains"
+        ],
+        "x-ms-request-id": [
+          "a36fd2e1e4307ff59e67a5216cbd64a6"
+        ],
+        "Cache-Control": [
+          "no-cache"
+        ],
+        "Date": [
+          "Thu, 23 Mar 2017 12:39:33 GMT"
+        ],
+        "Server": [
+          "1.0.6190.8255",
+          "(rd_rdfe_n.170217-1352)",
+          "Microsoft-HTTPAPI/2.0"
+        ]
+      },
+      "StatusCode": 200
+    }
+  ],
+  "Names": {},
+  "Variables": {
+    "SubscriptionId": "7a8c4b2b-1343-4ead-903f-031163054294"
+  }
+}
