+++ src/ServiceManagement/RecoveryServicesVaultUpgrade/VaultUpgrade.Tests/VaultUpgrade.Tests.csproj
@@ -1,0 +1,51 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <ProjectGuid>{5BC5A56E-229B-416B-9F9A-AEC777FF506C}</ProjectGuid>
+    <RootNamespace>VaultUpgrade.Tests</RootNamespace>
+    <AssemblyName>VaultUpgrade.Tests</AssemblyName>
+    <NuGetPackageImportStamp>
+    </NuGetPackageImportStamp>
+  </PropertyGroup>
+  <Import Project="..\..\..\..\tools\Library.Settings.targets" />
+  <Import Project="$(LibraryToolsFolder)\Test.Dependencies.target" />
+  <ItemGroup>
+    <Compile Include="Helpers\RecordedDelegationHandler.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ScenarioTests\VaultUpgradeTests.cs" />
+    <Compile Include="ScenarioTests\VaultUpgradeTestsBase.cs" />
+    <Compile Include="Spec.TechSupport.VaultUpgrade\VaultUpgradeManagementTestUtilities.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="App.config">
+      <SubType>Designer</SubType>
+    </None>
+    <None Include="SessionRecords\VaultUpgrade.Tests.VaultUpgradeTests\VaultUpgrade_CheckPrereqsTests.json" />
+    <None Include="SessionRecords\VaultUpgrade.Tests.VaultUpgradeTests\VaultUpgrade_TrackResourceUpgradationTests.json" />
+    <None Include="SessionRecords\VaultUpgrade.Tests.VaultUpgradeTests\VaultUpgrade_UpgradeResourceTests.json" />
+    <None Include="VaultUpgrade.playlist">
+      <SubType>Designer</SubType>
+    </None>
+  </ItemGroup>
+  <ItemGroup>
+    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+  </ItemGroup>
+  <ItemGroup>
+    <Reference Include="Microsoft.Threading.Tasks, Version=1.0.12.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\..\..\packages\Microsoft.Bcl.Async.1.0.168\lib\net40\Microsoft.Threading.Tasks.dll</HintPath>
+    </Reference>
+    <Reference Include="System.Runtime.Serialization" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\TestDependencies\TestDependencies.csproj">
+      <Project>{40f35645-00ee-4df2-b66e-7624546b66df}</Project>
+      <Name>TestDependencies</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\VaultUpgradeManagement\VaultUpgradeManagement.csproj">
+      <Project>{7c9dec77-95ce-4e17-b4e3-d50176be39e9}</Project>
+      <Name>VaultUpgradeManagement</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup />
+</Project>
