+++ src/ServiceManagement/RecoveryServicesVaultUpgrade/VaultUpgradeManagement/Microsoft.WindowsAzure.Management.RecoveryServicesVaultUpgrade.nuget.proj
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <!--
+    Microsoft.Azure.Management.RecoveryServicesVaultUpgrade
+    -->
+    <SdkNuGetPackage Include="Microsoft.WindowsAzure.Management.RecoveryServicesVaultUpgrade">
+      <PackageVersion>1.0.0-preview</PackageVersion>
+      <Folder>$(MSBuildThisFileDirectory)</Folder>
+    </SdkNuGetPackage>
+  </ItemGroup>
+</Project>
