+++ src/ServiceManagement/RecoveryServicesVaultUpgrade/VaultUpgradeManagement/Properties/AssemblyInfo.cs
