--- src/ServiceManagement/VaultUpgrade/VaultUpgrade.Tests/SessionRecords/VaultUpgrade.Tests.VaultUpgradeTests/VaultUpgrade_UpgradeResourceTests.json
@@ -1,73 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "/7a8c4b2b-1343-4ead-903f-031163054294/cloudservices/RecoveryServices-DKIEGNVV3OAPUDVOTINGUOLCIJIDA743FNSWQAM4O4NEQUVZFBWA-West-US/resources/HRMBvtd2ToDogfood/~/HyperVRecoveryManagerVault/TestVault1/upgrade",
-      "EncodedRequestUri": "LzdhOGM0YjJiLTEzNDMtNGVhZC05MDNmLTAzMTE2MzA1NDI5NC9jbG91ZHNlcnZpY2VzL1JlY292ZXJ5U2VydmljZXMtREtJRUdOVlYzT0FQVURWT1RJTkdVT0xDSUpJREE3NDNGTlNXUUFNNE80TkVRVVZaRkJXQS1XZXN0LVVTL3Jlc291cmNlcy9IUk1CdnRkMlRvRG9nZm9vZC9+L0h5cGVyVlJlY292ZXJ5TWFuYWdlclZhdWx0L1Rlc3RWYXVsdDEvdXBncmFkZQ==",
-      "RequestMethod": "POST",
-      "RequestBody": "<ResourceUpgradeInput xmlns=\"http://schemas.microsoft.com/windowsazure\">\r\n  <NewResourcePath>/subscriptions/7a8c4b2b-1343-4ead-903f-031163054294/resourceGroups/TargetRg/providers/Microsoft.RecoveryServicesBVTD2/Vaults/TestVault1</NewResourcePath>\r\n</ResourceUpgradeInput>",
-      "RequestHeaders": {
-        "Content-Type": [
-          "application/xml"
-        ],
-        "Content-Length": [
-          "271"
-        ],
-        "Accept": [
-          "application/xml"
-        ],
-        "x-ms-client-request-id": [
-          "8ea38d08-ff70-43d4-8398-c348c4c187bd"
-        ],
-        "x-ms-version": [
-          "2013-03-01"
-        ],
-        "User-Agent": [
-          "Microsoft.WindowsAzure.Management.VaultUpgrade.VaultUpgradeManagementClient/1.0.0.0"
-        ]
-      },
-      "ResponseBody": "<ResourceUpgradeDetails xmlns=\"http://schemas.microsoft.com/windowsazure\" xmlns:i=\"http://www.w3.org/2001/XMLSchema-instance\">\r\n  <EndTimeUtc>9999-12-31T23:59:59.9999999</EndTimeUtc>\r\n  <LastUpdatedTimeUtc>2017-03-16T08:03:50.9301757Z</LastUpdatedTimeUtc>\r\n  <Message></Message>\r\n  <OperationId>8ea38d08-ff70-43d4-8398-c348c4c187bd</OperationId>\r\n  <PreviousResourceId i:nil=\"true\" />\r\n  <StartTimeUtc>2017-03-16T08:03:50.9301757Z</StartTimeUtc>\r\n  <Status>InProgress</Status>\r\n  <TriggerType>UserTriggered</TriggerType>\r\n  <UpgradedResourceId>/subscriptions/7a8c4b2b-1343-4ead-903f-031163054294/resourceGroups/TargetRg/providers/Microsoft.RecoveryServicesBVTD2/Vaults/TestVault1</UpgradedResourceId>\r\n</ResourceUpgradeDetails>",
-      "ResponseHeaders": {
-        "Content-Length": [
-          "688"
-        ],
-        "Content-Type": [
-          "application/xml; charset=utf-8"
-        ],
-        "Expires": [
-          "-1"
-        ],
-        "Pragma": [
-          "no-cache"
-        ],
-        "x-ms-servedbyregion": [
-          "uswest"
-        ],
-        "x-ms-client-request-id": [
-          "8ea38d08-ff70-43d4-8398-c348c4c187bd"
-        ],
-        "Strict-Transport-Security": [
-          "max-age=15724800; includeSubDomains"
-        ],
-        "x-ms-request-id": [
-          "1475daa092a97fe5bf6f217394fba861"
-        ],
-        "Cache-Control": [
-          "no-cache"
-        ],
-        "Date": [
-          "Thu, 16 Mar 2017 08:03:51 GMT"
-        ],
-        "Server": [
-          "1.0.6190.8255",
-          "(rd_rdfe_n.170217-1352)",
-          "Microsoft-HTTPAPI/2.0"
-        ]
-      },
-      "StatusCode": 200
-    }
-  ],
-  "Names": {},
-  "Variables": {
-    "SubscriptionId": "7a8c4b2b-1343-4ead-903f-031163054294"
-  }
-}
