--- src/ServiceManagement/VaultUpgrade/VaultUpgrade.sln
@@ -1,54 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestDependencies", "..\..\TestDependencies\TestDependencies.csproj", "{40F35645-00EE-4DF2-B66E-7624546B66DF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VaultUpgradeManagement", "VaultUpgradeManagement\VaultUpgradeManagement.csproj", "{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VaultUpgrade.Tests", "VaultUpgrade.Tests\VaultUpgrade.Tests.csproj", "{5BC5A56E-229B-416B-9F9A-AEC777FF506C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Net40-Debug|Any CPU = Net40-Debug|Any CPU
-		Net40-Release|Any CPU = Net40-Release|Any CPU
-		Net45-Debug|Any CPU = Net45-Debug|Any CPU
-		Net45-Release|Any CPU = Net45-Release|Any CPU
-		Portable-Debug|Any CPU = Portable-Debug|Any CPU
-		Portable-Release|Any CPU = Portable-Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{40F35645-00EE-4DF2-B66E-7624546B66DF}.Net40-Debug|Any CPU.ActiveCfg = Net40-Debug|Any CPU
-		{40F35645-00EE-4DF2-B66E-7624546B66DF}.Net40-Release|Any CPU.ActiveCfg = Net40-Release|Any CPU
-		{40F35645-00EE-4DF2-B66E-7624546B66DF}.Net45-Debug|Any CPU.ActiveCfg = Net45-Debug|Any CPU
-		{40F35645-00EE-4DF2-B66E-7624546B66DF}.Net45-Debug|Any CPU.Build.0 = Net45-Debug|Any CPU
-		{40F35645-00EE-4DF2-B66E-7624546B66DF}.Net45-Release|Any CPU.ActiveCfg = Net45-Release|Any CPU
-		{40F35645-00EE-4DF2-B66E-7624546B66DF}.Net45-Release|Any CPU.Build.0 = Net45-Release|Any CPU
-		{40F35645-00EE-4DF2-B66E-7624546B66DF}.Portable-Debug|Any CPU.ActiveCfg = Portable-Debug|Any CPU
-		{40F35645-00EE-4DF2-B66E-7624546B66DF}.Portable-Release|Any CPU.ActiveCfg = Portable-Release|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Net40-Debug|Any CPU.ActiveCfg = Net40-Debug|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Net40-Debug|Any CPU.Build.0 = Net40-Debug|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Net40-Release|Any CPU.ActiveCfg = Net40-Release|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Net40-Release|Any CPU.Build.0 = Net40-Release|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Net45-Debug|Any CPU.ActiveCfg = Net45-Debug|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Net45-Debug|Any CPU.Build.0 = Net45-Debug|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Net45-Release|Any CPU.ActiveCfg = Net45-Release|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Net45-Release|Any CPU.Build.0 = Net45-Release|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Portable-Debug|Any CPU.ActiveCfg = Portable-Debug|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Portable-Debug|Any CPU.Build.0 = Portable-Debug|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Portable-Release|Any CPU.ActiveCfg = Portable-Release|Any CPU
-		{7C9DEC77-95CE-4E17-B4E3-D50176BE39E9}.Portable-Release|Any CPU.Build.0 = Portable-Release|Any CPU
-		{5BC5A56E-229B-416B-9F9A-AEC777FF506C}.Net40-Debug|Any CPU.ActiveCfg = Net40-Debug|Any CPU
-		{5BC5A56E-229B-416B-9F9A-AEC777FF506C}.Net40-Release|Any CPU.ActiveCfg = Net40-Release|Any CPU
-		{5BC5A56E-229B-416B-9F9A-AEC777FF506C}.Net45-Debug|Any CPU.ActiveCfg = Net45-Debug|Any CPU
-		{5BC5A56E-229B-416B-9F9A-AEC777FF506C}.Net45-Debug|Any CPU.Build.0 = Net45-Debug|Any CPU
-		{5BC5A56E-229B-416B-9F9A-AEC777FF506C}.Net45-Release|Any CPU.ActiveCfg = Net45-Release|Any CPU
-		{5BC5A56E-229B-416B-9F9A-AEC777FF506C}.Net45-Release|Any CPU.Build.0 = Net45-Release|Any CPU
-		{5BC5A56E-229B-416B-9F9A-AEC777FF506C}.Portable-Debug|Any CPU.ActiveCfg = Portable-Debug|Any CPU
-		{5BC5A56E-229B-416B-9F9A-AEC777FF506C}.Portable-Release|Any CPU.ActiveCfg = Portable-Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
