--- src/ServiceManagement/VaultUpgrade/VaultUpgradeManagement/Microsoft.WindowsAzure.Management.VaultUpgrade.nuget.proj
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <!--
-    Microsoft.Azure.Management.VaultUpgrade
-    -->
-    <SdkNuGetPackage Include="Microsoft.WindowsAzure.Management.VaultUpgrade">
-      <PackageVersion>1.0.0-preview</PackageVersion>
-      <Folder>$(MSBuildThisFileDirectory)</Folder>
-    </SdkNuGetPackage>
-  </ItemGroup>
-</Project>
